# lib/api-zod/src/generated/types/openaiConversation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Finora AI Financial Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiConversation {
  id: number;
  title: string;
  plan: string;
  createdAt: Date;
}