# lib/api-zod/src/generated/types/openaiConversationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Finora AI Financial Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiConversationInput {
  title: string;
  plan?: string;
}